Rename for better encaptulation
